move SST methods to separate file
